+++ ShopWebsite/Areas/BackEnd/Models/OrderProduct.cs
@@ -5,15 +5,15 @@ namespace ShopWebsite.Areas.BackEnd.Models;
 
 public partial class OrderProduct
 {
-    public int OrderProductId { get; set; }
+    public long OrderProductId { get; set; }
 
     public long OrderId { get; set; }
 
-    public int ProductId { get; set; }
+    public long ProductId { get; set; }
 
-    public int Quantity { get; set; }
+    public long Quantity { get; set; }
 
-    public int Price { get; set; }
+    public long Price { get; set; }
 
-    public int? Discount { get; set; }
+    public long? Discount { get; set; }
 }
