+++ ShopWebsite/Areas/BackEnd/Models/Product.cs
@@ -35,7 +35,7 @@ public partial class Product
 
     public string? ProductVideo1 { get; set; }
 
-    public int ProductPrice { get; set; }
+    public long ProductPrice { get; set; }
 
     public bool? ProductPublish { get; set; }
